requirements.txt
# requirements.txt

Flask>=2.0
Flask-SQLAlchemy>=2.5
jaconv
torch>=1.9
unidic-lite
mecab-python3
numpy
pandas
